% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.thresholds.r
\name{calculate.thresholds}
\alias{calculate.thresholds}
\title{Calculate the thresholds associated with a given truncation rule and given facies proportions when the number of GRFs are independent.}
\usage{
calculate.thresholds(flag, nthres, proportions, ymin, ymax)
}
\arguments{
\item{flag}{A vector with category numbers codifying the truncation rule.}

\item{nthres}{A vector with the number of thresholds for each GRF (1 x nfield).}

\item{proportions}{A vector with the proportions for each facies.}

\item{ymin}{A value representing negative infinity (-Inf). The 'ymin' value is updated recursively when the function is used iteratively.}

\item{ymax}{A value representing positive infinity (Inf). The 'ymax' value is updated recursively when the function is used iteratively.}
}
\value{
A vector with the calculated thresholds.
}
\description{
Calculate Thresholds for Truncation Rule and Facies Proportions
}
\details{
Use calculate_thresholds(flag, nthres, proportions) and do not define the arguments 'ymin' and 'ymax', which are used recursively.
}
\examples{
# Example usage
calculate.thresholds(flag = c(2,1,3,2,1,5,2,4,4), nthres = c(2,2), 
proportions = c(0.426, 0.253, 0.047, 0.205, 0.069))
}
\references{
This function is a translation of MATLAB code written by Emery (2007).
- Emery, X. (2007). Simulation of geological domains using the plurigaussian model: New developments and computer programs. Computers & Geosciences, 33(9), 1189–1201. DOI: 10.1016/j.cageo.2007.01.006
}
