% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facies.contacts.r
\name{facies.contacts}
\alias{facies.contacts}
\title{Calculate facies contacts}
\usage{
facies.contacts(data_Facies, facies_color, display)
}
\arguments{
\item{data_Facies}{A data frame with columns 'id' and 'Facies', where 'id' determines the borehole ID and 'Facies' contains the facies categories. 
If no 'id' column is defined, insert the 'x' and 'y' columns instead.}

\item{facies_color}{A vector with colors representing each facies category, arranged in alphabetical order. This argument is used when the contact probabilities 'prot' are produced (if not equal to NULL).}

\item{display}{A value indicating whether to produce bar plots of contacts for each facies (1) or not (0).}
}
\value{
The frequencies of facies contacts and their respective plots.
}
\description{
This function calculates the contacts of facies in a borehole dataset and plots them.
}
\examples{
data_Facies <- SFM_data
facies_color <- c("orange", "black", "red", "green", "blue") 
# Colors in alphabetical order of facies: AL, CO, HD, MG, SG
facies.contacts(data_Facies, facies_color, display = 1)

}
