% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.scores.r
\name{normal.scores}
\alias{normal.scores}
\title{Transform a variable into Normal Scores}
\usage{
normal.scores(values)
}
\arguments{
\item{values}{A vector of values to be transformed.}
}
\description{
Transform a variable into Normal Scores
}
\details{
The z values weights are taken as equal
}
