% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.values.r
\name{replace.values}
\alias{replace.values}
\title{Replace values}
\usage{
replace.values(input_data, old_values, new_values)
}
\arguments{
\item{input_data}{A vector of values (N) to be replaced.}

\item{old_values}{A vector of length n containing the values to be replaced.}

\item{new_values}{A vector of length n containing the values to replace the old_values.}
}
\description{
Replace the values of a categorical variable with new values
}
\note{
The length of old_values and new_values should be equal.
}
