% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{SFM_data}
\alias{SFM_data}
\title{Samples of 639 facies}
\format{
A data frame with 639 rows and 4 variables:
\describe{
  \item{x}{coordinates in m along East-West }
  \item{y}{coordinates in m along North-South}
  \item{z}{coordinates in m along vertical direction}
  \item{Facies}{facies category}
}
}
\usage{
SFM_data
}
\description{
A dataset containing information on 639 lithofacies samples from 38 boreholes in a lignite mine in Greece.
The facies are grouped into five categories: clay (AL), hard rock (HD), marl (MG), sand with gravels (SG), and coal (CO).
}
\keyword{datasets}
