% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.r
\name{vpc}
\alias{vpc}
\title{Vertical Proportions Curves}
\usage{
vpc(data_Facies, vpc_step, xy_grid)
}
\arguments{
\item{data_Facies}{A data frame with columns x, y, z coordinates, and the Facies data. The columns must have the names x, y, z and Facies, respectively.
Facies colum contains the category of Facies in alphabetic code or numeric code}

\item{vpc_step}{A list of x, y, and z wich define the partition of the examined gird}

\item{xy_grid}{A list of arguments to define the grid parameters. Use list(xmin, xmax, ymin,ymax),
where: xmin, xmax, ymin, ymax are single values indicating the minimum and maximum grid coordinates along EW and NS directions,
where the proportions curves are calculated}
}
\value{
An array of proportions for each partiton of the grid (array)
}
\description{
Calculate proportion curves of faces along the vertical direction over the domain in a grid (also global proportion curves).
}
\examples{
vpc_step<-list(x = 2000, y = 2000, z=20)
xy_grid <-list(xmin=-16000, xmax = -12000, ymin = 16000, ymax = 20000)
vpc_results <- vpc(SFM_data, vpc_step, xy_grid)
## Example to plot the VPC using ggplot2 package
\dontrun{

library("ggplot2")
depth_scale <- as.numeric(dimnames(vpc_results)[[1]])
facies_categories <- dimnames(vpc_results)[[2]]
plot_title <- dimnames(vpc_results)[[3]]
ndepth_intervals <- nrow(vpc_results)
proportions_categories <- sort(rep(facies_categories, ndepth_intervals))
facies_colors <- c("orange", "black", "red","green","blue")
for (i in 1:dim(vpc_results)[[3]]){
proportions_vector <- c(vpc_results[ , ,i])
proportions_data <- data.frame(depth_scale,proportions_vector,proportions_categories)
colnames(proportions_data) <- c("Depth","Proportions","Facies")
dev.new()
print(ggplot(proportions_data, aes(y = Proportions,x = Depth, fill = Facies)) +
geom_area(size = 1) +
scale_fill_manual(values = facies_colors) +
coord_flip() +
ggtitle(plot_title[i])+
theme(legend.position = "none"))}
}

}
