% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.thresholds.r
\name{calculate.thresholds}
\alias{calculate.thresholds}
\title{Calculate the thresholds associated with a given truncation rule and given facies proportions when the number of GRFs are independent.}
\usage{
calculate.thresholds(flag, nthres, proportions, ymin, ymax)
}
\arguments{
\item{flag}{is a vector with category numbers codifying the truncation rule}

\item{nthres}{is a vector with the number of thresholds for each GRF (1  x nfield)}

\item{proportions}{is a vector with the proportions for each facies}

\item{ymin}{a value equals to -Inf. ymin is updated whren the function is used recursively}

\item{ymax}{a value equals to Inf. ymax is updated whren the function is used recursively}
}
\value{
A vector with the thresholds
}
\description{
Calculate the thresholds associated with a given truncation rule and given facies proportions when the number of GRFs are independent.
}
\details{
Use calculate_thresholds(flag,nthres,proportions) and do not define the arguments ymin and ymax that are used recursively
}
