% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co.sim.r
\name{co.sim}
\alias{co.sim}
\title{Co-simulation}
\usage{
co.sim(
  simu_grid,
  conditioning_data,
  truncation_rule,
  variog_model,
  neigb_par,
  simu_par
)
}
\arguments{
\item{simu_grid}{A list of arguments to define the grid parameters. Use list(x0,y0,z0,nx,ny,nz,dx,dy,dz),
where: x0, y0, z0 are    single values indicating the minimum grid coordinates along EW, NS and vertical directions,
nx, ny, nz    are single values indicating the number of grid nodes along east, north and vertical directions,
dx, dy, dz    are single values indicating the grid meshes along EW, NS and vertical directions}

\item{conditioning_data}{A matrix of conditioning (categorical and continuous ) data (number of data x 6). The first columns contain the coordinates (x,y,z) of the conditioning data; the fourth column contains the conditioning values of the categorical data (numerical codes of facies). 
The fifth column contains the conditioning values of the continuous data and the last column contains the normal scores of the continuous data. Use NaN for the values in 4th to 6th column when categorical and continuous values are not at the same locations.
When no conditional data are imported, the argument remains empty, so an unconditional model is built.}

\item{truncation_rule}{A list of arguments to determine the truncation rule of Plurigaussian. Use list(nfield,nthres,thresholds,flag).
where: nfield is the number of GRFs used for the truncation rule,
nthres is a vector with the number of thresholds for each GRF (1  x nfield),
thresholds is a vector of the thresholds for all GRFs (1 x sum(nthres)),
flag is a vector with category numbers codifying the truncation rule}

\item{variog_model}{A list of arguments to define the nested variogram theoretical models. Use list(model,cc,b,nugget)
where: model is a matrix containing the covariance model for the GRFs (nested structures x 7 matrix). Each row corresponds to a nested structure and is codified as: type, scale factors, angles. Use the codes of the availiable types of variogram model (see details). There are three scale factors (along the rotated NS, EW and vertical axes) and three angles to define the coordinates rotation (azimuth, dip and plunge), see Deutsch & Journel (1997, p.25).
cc is a matrix indicating the sills of nested structures (nested structures x (1 + nfield)^2),
b is column vector with the additional parameters required for specific covariance types  (nested structures x 1), Also, see details.
nugget    is a row vector with nugget effect variance-covariance matrix with size of (1 + nfield)^2.}

\item{neigb_par}{A list of arguments to define the moving neighborhood to condition the data. Use list list(radius,angles,octant,ndata) 
where: radius is a row vector row vector with the maximum search radii along the rotated NS, EW and vertical axes (1 x 3) for conditioning data,
angles is a row vector with the angles for anisotropic search (1 x 3), based on Deutsch and Journel (1997, p. 27),
octant takes values 1 or 0, specifying if the neighborhood should be divided into octants (1) or not (0),
ndata is a single value with the number of conditioning data per octant or in total}

\item{simu_par}{A list of arguments to define the simulation parameters. Use list(nlines,nrealiz,seed,nnodes,itGibbs),
where: nlines is a single value indicating the number of lines to use for simulating the nested structures by turning bands,
nrealiz is a single value indicating the number of realizations to produce,
seed is the seed number for generating random values}
}
\value{
A list of the simulated coordinates and the respective simulated categorical and continuous values.
coord contains the center of gridded blocks (number of blocks x 3), 
categoricalVar contains the simulated values of categorical variable (number of blocks x nrealiz). Each column is a realization of categorical variable.  
continuousVar contains the simulated values of continuous variable (number of blocks x nrealiz)
}
\description{
Conditional co-simulation of a continuous (transformed multi-Gaussian) and categorical (plurigaussian) GRFs using the turning bands method.
The function simulates categorical (facies) and continuous variables using spatially correlated GRFs. The GRFs  are stationary and jointly Gaussian,
with zero means and unit variances. The GRFs should be correlated.
Multiple realizations of the simulated facies can be produced in a regular grid.
}
\seealso{
pluri.sim
}
