% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facies.contacts.r
\name{facies.contacts}
\alias{facies.contacts}
\title{Calculate facies contacts}
\usage{
facies.contacts(data_Facies, facies_color, display)
}
\arguments{
\item{data_Facies}{A data frame having colums with names id and Facies. id colum determines the id of a borehole and Facies colum contains the category of Facies. Insert the x and y columns in the case no id coulum is defined.}

\item{facies_color}{A vector with colors (color for each category) in alphabetical order of facies. The argument is used when the contacts probabilities prot are produced (if not = NULL).}

\item{display}{A value indicates to produce (=1) or not (=0) the barplots of contacts for each facies.}
}
\value{
The frequencies of facies contacts and the respective plots
}
\description{
Calculate the contacts of facies in a borehole dataset and plot them
}
\examples{
data_Facies <- SFM_data
facies_color<-c("orange", "black", "red","green","blue") 
# color in facies Alphabetic order (AL,CO,HD,MG,SG)
display <- 1

}
