% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.scores.r
\name{normal.scores}
\alias{normal.scores}
\title{Transform a Variable into Normal Scores}
\usage{
normal.scores(values)
}
\arguments{
\item{values}{A vector of values to be transformed.}
}
\description{
This function transforms a variable into normal scores.
}
\details{
The z-values weights are taken as equal.
Ensure that the provided data values follow a normal distribution for accurate transformation into normal scores.
}
