% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvmodel.r
\name{rvmodel}
\alias{rvmodel}
\title{Modeling Indicator Variograms}
\usage{
rvmodel(truncation_rule, variog_model, h_lag, nlag, azm, dip, nMC)
}
\arguments{
\item{truncation_rule}{A list of arguments to determine the truncation rule of Plurigaussian. Use list(nfield,nthres,thresholds,flag).
where: nfield is the number of GRFs used for the truncation rule,
nthres is a vector with the number of thresholds for each GRF (1  x nfield),
thresholds is a vector of the thresholds for all GRFs (1 x sum(nthres)),
flag is a vector with category numbers codifying the truncation rule}

\item{variog_model}{A list of arguments to define the nested variogram theoretical models. Use list(model,cc,b,nugget)
where: model is a matrix containing the covariance model for the GRFs (nested structures x 7 matrix). Each row corresponds to a nested structure and is codified as: type, scale factors, angles. Use the codes of the availiable types of variogram model (see details). There are three scale factors (along the rotated NS, EW and vertical axes) and three angles to define the coordinates rotation (azimuth, dip and plunge), see Deutsch & Journel (1997, p.25).
cc is a matrix indicating the sills of nested structures (nested structures x nfield^2). In the case of co-simulation modeling, cc is a matrix with dimension (nested structures x (1 + nfield)^2)
b is column vector with the additional parameters required for specific covariance types  (nested structures x 1), Also, see details.
nugget is a row vector with nugget effect variance-covariance matrix with size of nfield^2. In the case of co-simulation modeling, nugget is a row vector with dimension (1 + nfield)^2.}

\item{h_lag}{The lag distance (single value)}

\item{nlag}{The number of lags (single value)}

\item{azm}{azimuth (single value)}

\item{dip}{dip (single value)}

\item{nMC}{The number of Monte Carlo realizations (single value)}
}
\value{
An array of semi-variograms between rock type indicators (ncategory * ncategory * nlag matrix). In the case of co-simulation modeling the function returns a list including:
pg_variograms: semi-variograms between facies type indicators (ncategory * ncategory * nlag matrix)
cross_variog: semi-variograms between Gaussian field Z0 and facies type indicators (ncategory * nlag matrix)
}
\description{
Variogram (simple and cross) Modeling of Facies Indicator or/and cross semi-variograms between Z0 and Facies indicator (Co-simulation) for the specified direction and lag distances.
The function helps to determine the model parameters, in particular the threshold values and the simple and cross semi-variograms of the GRFs.
The calculation of simple and cross-variograms is performed by Monte Carlo simulation.
}
\details{
The output of the function is a list of two matrices: $coord contains the center of blocks in the grid (number of blocks x 3) and $categoricalVar contains the simulated values of categorical variable (number of blocks x nrealiz - each column is a realization of the categorical variable).
Available types of variogram model:
1: spherical
2: exponential (beware that range = 3*scale factor)
3: gamma (parameter b > 0)
4: stable (parameter b between 0 and 2)
5: cubic
6: Gaussian
7: cardinal sine
8: Bessel J (parameter b > 0)
9: Bessel K (parameter b > 0)
10: generalized Cauchy (parameter b)
11: exponential sine
}
\examples{
truncation_rule <- list(nfield = 2 , nthres = c(2,2),thresholds = 
c(-0.5769, 1.0762, -0.4895, 0.6592), flag = c(2,1,3,2,1,5,2,4,4))
# Gaussian Random Fields Modeling
nst_str <- 2
nugget <- c(0,0,0,0)
B_nst1 <- matrix(c(0.9986,-0.1,-0.1,0.8167),nst_str,truncation_rule$nfield)
B_nst2 <- matrix(c(0.00149,0,0,0.1833),nst_str,truncation_rule$nfield)
sills_matrix <- matrix(c(B_nst1,B_nst2),nrow=nst_str,ncol=truncation_rule$nfield^2,byrow=TRUE)
variogram_models <- matrix(c(1,400,400,140,0,0,0,2,300,300,90,0,0,0),nst_str,7,byrow=TRUE)			
variog_model <- list(model=variogram_models,cc=sills_matrix,b=matrix(0,nrow=nst_str,ncol=1),
nugget=nugget)
azm <- 0
dip <- 0
h_lag <- 100
nlag <- 10
nMC <- 1e4;
}
