% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.r
\name{vpc}
\alias{vpc}
\title{Calculate Vertical Proportions Matrix}
\usage{
vpc(data_Facies, vpc_step, xy_grid)
}
\arguments{
\item{data_Facies}{A data frame containing the columns 'x', 'y', 'z', and 'Facies' representing the coordinates and facies data. 
The column names must be exactly 'x', 'y', 'z', and 'Facies', respectively. 
The 'Facies' column should contain the facies categories in alphabetical or numeric codes.}

\item{vpc_step}{A list of 'x', 'y', and 'z' values defining the partition of the examined grid.}

\item{xy_grid}{A list specifying the grid parameters. Use the format list(xmin, xmax, ymin, ymax) to indicate the minimum and maximum grid coordinates. 
along the east-west (EW) and north-south (NS) directions, where the proportion matrix is calculated.}
}
\value{
An array of proportions of facies for each partition of the grid (array).

details The function calculates the vertical proportion curves, which represent the proportion of each facies category along the vertical direction within each partition of the grid. It takes into account the provided facies data and their corresponding spatial coordinates.
}
\description{
This function calculates the proportion matrix of facies along the vertical direction over the domain in a grid.
}
\examples{
# Example usage
vpc_step<-list(x = 2000, y = 2000, z=20)
xy_grid <-list(xmin=-16000, xmax = -12000, ymin = 16000, ymax = 20000)
vpc_results <- vpc(SFM_data, vpc_step, xy_grid)
# Example to plot the Vertical Proportion Curves (VPC) using ggplot2 package
\dontrun{

library("ggplot2")
depth_scale <- as.numeric(dimnames(vpc_results)[[1]])
facies_categories <- dimnames(vpc_results)[[2]]
plot_title <- dimnames(vpc_results)[[3]]
ndepth_intervals <- nrow(vpc_results)
proportions_categories <- sort(rep(facies_categories, ndepth_intervals))
facies_colors <- c("orange", "black", "red","green","blue")
for (i in 1:dim(vpc_results)[[3]]){
proportions_vector <- c(vpc_results[ , ,i])
proportions_data <- data.frame(depth_scale,proportions_vector,proportions_categories)
colnames(proportions_data) <- c("Depth","Proportions","Facies")
dev.new()
print(ggplot(proportions_data, aes(y = Proportions,x = Depth, fill = Facies)) +
geom_area(size = 1) +
scale_fill_manual(values = facies_colors) +
coord_flip() +
ggtitle(plot_title[i])+
theme(legend.position = "none"))}
}

}
